requirements.txt
meshtastic
python-statemachine
